Avoid name collision between npc_simulator and scenario_api_utils
