+++ api/scenario_api_simulator/include/scenario_api_simulator/npc_route_manager.h
@@ -51,7 +51,7 @@ public:
   /**
    * @brief constructor
    */
-  NPCRouteManager(rclcpp::Node::SharedPtr node);
+  NPCRouteManager(rclcpp::Node& node);
 
   /**
    * @brief destructor
