+++ npc_simulator/package.xml
@@ -17,6 +17,7 @@
   <depend>dummy_perception_publisher</depend>
   <depend>lanelet2_extension</depend>
   <depend>rclcpp</depend>
+  <depend>scenario_api_utils</depend>
   <depend>sensor_msgs</depend>
   <depend>std_msgs</depend>
   <depend>tf2</depend>
